f64 usage fix

+++ assembly-line/src/lib.rs
@@ -1,7 +1,7 @@
 #![allow(unused)]
 
 pub fn production_rate_per_hour(speed: u8) -> f64 {
-    speed as f64 * 221 as f64 * success_rate(speed)
+    f64::from(speed) * 221.0 * success_rate(speed)
 }
 
 pub fn working_items_per_minute(speed: u8) -> u32 {
